source/requirements.txt
aws-cdk-lib>=2.140.0
boto3<2.0.0,>=1.24.11
aws-lambda-powertools
crhelper~=2.0.11
pytest
pytest-cov
pytest-env>=0.6.2
pytest-mock>=3.10.0
moto>=5.0.0
cdk-nag
docker
pyparsing
url_normalize
responses
pytest-ordering
-e cdk_solution_helper_py/helpers_cdk
-e cdk_solution_helper_py/helpers_common
-e infrastructure
-e tests
